requirements.txt
google-genai
pytest-html
python-dotenv